requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers for CLIP
transformers>=4.30.0
ftfy
regex

# Segmentation Models
segmentation-models-pytorch>=0.3.3

# Data Processing
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# Dataset Management
roboflow>=1.1.0

# Metrics and Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Logging and Experiment Tracking
tensorboard>=2.13.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
